add tentative support for "pre-baked" subtitles
It seems that, as of November 2025, Font.cpp doesn't properly support
superimposing diacritic-codepoints over alphabet-character-codepoints.

I briefly looked into asigning diacrtic codepoints negative character
widths to force them to align with the previous character, but knew
that kind of hack would break if diacritic aligning was properly
supported in the future.

For now, I've added support for "pre-baked" subtitles in the form of
sprite sheets.